fix(android): remove chat clean focus mode
